+++ spec/bali/components/avatar_spec.rb
@@ -25,4 +25,16 @@ RSpec.describe Bali::Avatar::Component, type: :component do
     expect(rendered_component).to have_css '.control > input[accept=".jpg, .jpeg, .png,"]'
     expect(rendered_component).to have_css '.control > input[name="test"]'
   end
+
+  it 'renders the avatar component with an image loaded' do
+    helper.form_with(url: '/') do |form|
+      options.merge!(method: :test, form: form, placeholder_url: '/bulma-default.png')
+      render_inline(component) do |c|
+        c.picture(image_url: '/avatar.png', class: 'test-image')
+      end
+    end
+
+    expect(rendered_component).to have_css '.avatar-component > figure.image'
+    expect(rendered_component).to have_css 'figure.image > img[src*="avatar.png"]'
+  end
 end