+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.25.0] - 2022-07-05
 
-- Add `Avatar::Component`. With this compoent we'll be able to see a preview of the image we want as an avatar.
+- Add `Avatar::Component`. With this component we'll be able to see a preview of the image we want as an avatar.
 
 ## [0.24.3] - 2022-07-05